requirements.txt
asgiref==3.2.3
boto3==1.11.7
botocore==1.14.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
django-filter==2.2.0
django-storages==1.8
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
jmespath==0.9.4
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.22.0
s3transfer==0.3.1
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
